requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pydantic
faiss-cpu 
langchain 
sentence-transformers
axios
PyPDF2